React/07_HyDE RAG/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi>=0.129.0",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.8",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "uvicorn[standard]>=0.40.0",
]
